requirements.txt
albumentations==1.0
GPUtil==1.4.0
h5py==3.2.1
imageio==2.9.0
imbalanced-learn==0.7.0
javabridge==1.0.19
matplotlib==3.2.2
openslide-python==1.1.2
pandas==1.2.4
peewee==3.14.4
Pillow==8.0.0
pretrainedmodels==0.7.4
pytest==6.2.4
python-bioformats==4.0.4
scikit-image==0.16.2
scikit-learn==1.0
scikit-learn-extra==0.2.0
scipy==1.6.2
seaborn==0.11.1
tqdm==4.61.0
typer==0.3.2
visdom==0.1.8.9
umap-learn==0.5.1